have working modal - still needs styling
