+++ app/src/views/QuizQuestionView.jsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 import { CheckBox } from 'react-native-elements';
-import BadgeAcquisition from './BadgeAcquisition';
-import BadgeAcquisition1 from './BadgeAcquisition1';
-
+import BadgeModal from './BadgeModal';
 
 const styles = {
   wrapper: {
@@ -132,7 +130,7 @@ export default class QuizQuestionView extends React.Component {
         {/* Is this the last question and is a choice selected */}
         {isLastQuestion && isSelected ? (
           <View>
-            <BadgeAcquisition1 courseBadgeId={courseBadgeId} />
+            <BadgeModal courseBadgeId={courseBadgeId} />
 
             {/* <BadgeAcquisition courseBadgeId={courseBadgeId} /> */}
           </View>
